+++ src/features/auth/login/data/README.md
@@ -1,0 +1,15 @@
+# Voyager curated servers policy
+
+[servers.ts](./servers.ts) contains whitelisted and categorized instances.
+
+Please feel free to open an issue or PR to add an instance to, provided the instance meets the following criteria:
+
+1. Server uptime of at least 3 months
+2. At least 50 MAU, **UNLESS:**
+   1. a regional or niche category instance, or
+   2. added only to `ADDITIONAL_LOGIN_INSTANCES` (only shown for login, not signup)
+3. Commitment to provide 2 months lead time before shut down
+4. Active moderation against harassment, bullying, racism, discrimination, transphobia, hate speech, violation of privacy and threats of violence.
+5. Privacy Policy and Terms of Use set up on `/legal` page
+
+The above policy can be bypassed for instances added by me (aeharding). I reserve the right to refuse to add an instance for any reason without explanation.
